refactor: read-context-files 移入 session、time-perception 实现移入 capability 目录，context/ 收敛为 context window 管理域
